php-stan-update (#654)
* test: aumentando level php stan para 3

* test: aumentando level php stan para 5

* test: aumentando level php stan para 5

* Update GenerateMfpKey.php

* Update CalendarTools.php

+++ app/DTO/InvoiceItemDTO.php
@@ -87,7 +87,7 @@ class InvoiceItemDTO
 
     public function getNextInstallmentMonth(): int
     {
-        return substr($this->nextInstallment, 5, 2);
+        return (int)substr($this->nextInstallment, 5, 2);
     }
 
     protected function setNextInstallment(string $nextInstallment): void
